refactor: 설정 하는 부분을 별도의 객체로 이동
* refactor: GameExecutor 초기화 로직을 Configuration으로 이동
* refactor: 패키지 이동
